rename reset -> replay on dashboard
