+++ src/Beckett.Dashboard/Subscriptions/Routes.cs
@@ -1,5 +1,5 @@
 using Beckett.Dashboard.Subscriptions.Pause;
-using Beckett.Dashboard.Subscriptions.Reset;
+using Beckett.Dashboard.Subscriptions.Replay;
 using Beckett.Dashboard.Subscriptions.Resume;
 using Beckett.Dashboard.Subscriptions.Subscription;
 using Beckett.Dashboard.Subscriptions.Subscriptions;
@@ -15,7 +15,7 @@ public class Routes : IConfigureRoutes
         routes.MapGet("/", SubscriptionsEndpoint.Handle);
         routes.MapGet("/{groupName}/{name}", SubscriptionEndpoint.Handle);
         routes.MapPost("/{groupName}/{name}/pause", PauseEndpoint.Handle).DisableAntiforgery();
-        routes.MapPost("/{groupName}/{name}/reset", ResetEndpoint.Handle).DisableAntiforgery();
+        routes.MapPost("/{groupName}/{name}/replay", ReplayEndpoint.Handle).DisableAntiforgery();
         routes.MapPost("/{groupName}/{name}/resume", ResumeEndpoint.Handle).DisableAntiforgery();
     }
 }
